fix(cli): match PATH bin dir as a whole component; add changelog entry
rc_configures_path used a bare `contains(bin_str)` substring check, which false-positives on unrelated paths that merely have the bin dir as a prefix (e.g. `.astrid/bin_backup`, `.astrid/bin/sub`). That made the PATH guard skip the real `.astrid/bin` setup and silently leave astrid off PATH — the exact silent-skip failure mode init must avoid.

Now the guard returns already-configured only when the exact export block is present OR the bin dir appears as a WHOLE path-list token — bounded on BOTH sides by a shell PATH separator (start/end or one of : " ' = ( ) space tab newline/CR). A trailing / or alphanumeric/_/- is not a boundary, so bin_backup and bin/foo correctly fall through to adding the block. When unsure it errs toward adding: a duplicate PATH entry is harmless, a silent skip is not. Boundary offsets use saturating_add to satisfy the arithmetic_side_effects lint.

Added rc_path_guard_rejects_substring_false_positives (fails against the old substring check). Extracted self_update.rs tests to self_update_tests.rs to stay under the 1000-line file cap. Added a CHANGELOG [Unreleased] Fixed entry covering the per-principal init fix, provisioning honesty, and authenticated GitHub API resolution.
