+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Changelog tracking starts with 0.2.0. Prior versions were not tracked.
 
 ### Fixed
 
+- **`astrid init` now honours the global `--principal` flag, provisioning the named principal's home instead of always `default`.** `astrid init --principal <x>` (or `ASTRID_PRINCIPAL` / the active-agent context) previously wrote the distro's per-capsule env files and the `Distro.lock` under `default` even though the capsule files themselves already landed under the resolved principal — so a scoped principal was left with an empty capsule view, an empty-hash lock, and a re-provision on every run (its lock lived under the wrong home and never satisfied the principal-aware auto-init freshness check). Init now threads the resolved principal — the same identity `astrid capsule install` uses — through the lock path, env files, per-provider onboarding, and every post-install read-back, both on the online path and the offline `.shuttle` path. Provisioning is also honest: a run where every capsule failed no longer prints "Installation complete" or writes a `Distro.lock` (a stale lock would wedge re-runs) and exits non-zero; a partial run writes NO lock and reports the true installed/total count so a re-run retries the missing capsules instead of short-circuiting at the version-only freshness gate; the shell-PATH edit is idempotent via a whole-path-component match (never a bare substring, so it neither appends a duplicate block nor silently skips setup when an unrelated `…/bin_backup` path is present); and GitHub release/source resolution now authenticates with `GH_TOKEN` / `GITHUB_TOKEN` when present — lifting the anonymous 60/hr API ceiling that a full distro can exhaust mid-provision — while degrading gracefully to anonymous when no token is set. No WIT / host-ABI change — CLI provisioning only. Closes #1095.
 - **Restored shared-by-hash single-instance capsule loading, correcting the per-principal-instance regression that #1083 introduced against #1069's own spec.** #1083 keyed loaded runtime instances by `(principal, wasm_hash)` and built a **separate** `Arc<dyn Capsule>` runtime (and WASM build) per principal, so a capsule referenced by N principals loaded N times — memory scaled principals × capsules. The registry now keys instances by content hash alone: a hash referenced by N principals loads **exactly once** (one runtime, one WASM build), with the per-principal `name -> hash` views layered on top. A per-hash refcount tracks how many principal views reference the shared runtime; it is cancelled/unloaded only when the **last** view releases it (`unregister_for` now returns whether the instance was torn down, so kernel unload/restart never cancels a runtime other principals still use). The kernel load path dedups by hash: before building, a loaded hash short-circuits to a view-add via `register_existing`/`contains_hash` (no second runtime). **Kept from #1083:** per-principal dispatch/visibility views (`cloned_values_for` scoping, admin bypass, fail-closed view-scoped describe fan-out), per-principal on-disk install sets, and per-invocation env/KV/secret/home overlays. Cancellation granularity now matches the shared-instance model too: each invocation waits on a per-principal child token of the instance token (installed with the other per-invocation overlays), and a non-last view release calls a new `request_cancel_for(principal)` that cancels exactly the departing principal's in-flight blocking host calls (approval/elicit waits, net/io/ipc waits) — so a `capsule remove` with a pending approval can no longer wedge the shared runtime for every other principal, while a full unload still cancels everything via the instance token cascade. Closes #1069; corrects #1083.
 - **Closed the cross-principal host-state fallback that a shared runtime would otherwise open.** A shared instance is loaded under no real principal's identity, so its load-time host state is a NEUTRAL, fail-closed placeholder: the `kv` is a physically-isolated in-memory store and the `secret_store` is deny-all — never `default`'s (or anyone's) real KV, secrets, or home. The `effective_kv` / `effective_secret_store` / `effective_home` / `effective_tmp` / `effective_capsule_log` accessors therefore fall back to that neutral placeholder for a principal-less / load-time invocation (which denies rather than exposing any principal's state), while EVERY caller that carries a principal — the owner/`default` included — gets its own `invocation_*` overlay scoped to the invoking principal. A regression test constructs a shared instance and asserts a `bob`-stamped invocation reads/writes bob's KV and secret namespaces, an `alice`-stamped invocation lands in alice's, and a principal-less invocation hits only the neutral placeholder. Per-invocation env config is unaffected: it is read per call from the invoking principal's `.config/env/{capsule}.env.json`, not baked into the load-time instance. Relates to #977, #1069.
 - **Made the shared capsule runtime present one stable, content-addressed IPC `source_id` to every principal, and kept the gateway's reply-trust derivation byte-identical.** The WASM engine's `capsule_uuid` seed and the gateway's `capsule_source_id_v1` both drop the principal segment (now `{capsule_id}\0{content_hash}`, same fixed namespace), so a shared runtime's replies carry one id all principals accept. Cross-principal reply routing is unaffected because it already keys on the principal-scoped routed subscription plus the body correlation id — the `source_id` is only an authenticity gate. A forged reply (wrong capsule/hash) is still rejected, and a new lockstep test pins the two derivations together so they cannot drift. Relates to #1069.
